+++ src/odr/internal/ooxml/spreadsheet/ooxml_spreadsheet_document.hpp
@@ -24,8 +24,8 @@ public:
   [[nodiscard]] bool is_editable() const noexcept override;
   [[nodiscard]] bool is_savable(bool encrypted) const noexcept override;
 
-  void save(const Path &path) const override;
-  void save(const Path &path, const char *password) const override;
+  void save(std::ostream &out) const override;
+  void save(std::ostream &out, const char *password) const override;
 
 private:
   XmlDocumentsAndRelations m_xml_documents_and_relations;