+++ src/odr/internal/rtf/rtf_document.hpp
@@ -20,8 +20,8 @@ public:
   [[nodiscard]] bool is_editable() const noexcept override;
   [[nodiscard]] bool is_savable(bool encrypted) const noexcept override;
 
-  void save(const Path &path) const override;
-  void save(const Path &path, const char *password) const override;
+  void save(std::ostream &out) const override;
+  void save(std::ostream &out, const char *password) const override;
 
 private:
   ElementRegistry m_element_registry;